Rename schemas/profiles -> profile
