+++ imports/client/components/CallLinkBox.tsx
@@ -15,7 +15,7 @@ import CallSignals from '../../lib/models/call_signals';
 import Profiles from '../../lib/models/profiles';
 import { CallParticipantType } from '../../lib/schemas/call_participant';
 import { CallSignalType, CallSignalMessageType } from '../../lib/schemas/call_signal';
-import { ProfileType } from '../../lib/schemas/profiles';
+import { ProfileType } from '../../lib/schemas/profile';
 import { RTCConfigType } from '../rtc_config';
 import Spectrum, { SpectrumHandle } from './Spectrum';
