+++ imports/client/components/CallSection.tsx
@@ -15,7 +15,7 @@ import { getAvatarCdnUrl } from '../../lib/discord';
 import CallParticipants from '../../lib/models/call_participants';
 import Profiles from '../../lib/models/profiles';
 import { CallParticipantType } from '../../lib/schemas/call_participant';
-import { ProfileType } from '../../lib/schemas/profiles';
+import { ProfileType } from '../../lib/schemas/profile';
 import { RTCConfigType, RTCConfig } from '../rtc_config';
 import CallLinkBox from './CallLinkBox';
 import Spectrum from './Spectrum';
