+++ imports/lib/models/profiles.ts
@@ -1,5 +1,5 @@
 import { Meteor } from 'meteor/meteor';
-import ProfilesSchema, { ProfileType } from '../schemas/profiles';
+import ProfileSchema, { ProfileType } from '../schemas/profile';
 import Base from './base';
 
 const Profiles = new class extends Base<ProfileType> {
@@ -26,7 +26,7 @@ const Profiles = new class extends Base<ProfileType> {
     return displayNames;
   }
 }();
-Profiles.attachSchema(ProfilesSchema);
+Profiles.attachSchema(ProfileSchema);
 
 // Ideally, we'd only publish profiles whose hunt set overlaps with
 // yours, but that's hard so for now publish everything
